WIP: migrate foreground state management to separate concept module
