+++ kotlin/concepts/concept-foreground-state/src/main/java/io/matthewnelson/concept_foreground_state/ForegroundState.kt
@@ -13,7 +13,7 @@
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 * */
-package io.matthewnelson.feature_authentication_core.model
+package io.matthewnelson.concept_foreground_state
 
 sealed class ForegroundState {
     object Background: ForegroundState()
