+++ settings.gradle
@@ -26,7 +26,7 @@ include ':kotlin:common:build-config'
 include ':kotlin:concepts:authentication:concept-authentication'
 include ':kotlin:concepts:authentication:concept-encryption-key'
 include ':kotlin:concepts:concept-coroutines'
-include ':kotlin:condepts:concept-foreground-state'
+include ':kotlin:concepts:concept-foreground-state'
 include ':kotlin:concepts:concept-navigation'
 include ':kotlin:concepts:concept-views'
